% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gswdesign-package.R
\docType{package}
\name{gswdesign-package}
\alias{gswdesign-package}
\title{gswdesign: The Gram-Schmidt walk design}
\description{
Provides wrappers for the julia package 'gswdesign', which is a
computationally efficient implementation of the Gram-Schmidt walk design.
}
\details{
See \code{\link{sample_gsw}} for the main function.

See the package's website for more information:
\url{https://github.com/xxx/xxx}.

Bug reports and suggestions are greatly appreciated. They are best reported
here: \url{https://github.com/xxx/xxx/issues}.
}
\references{
Harshaw, Christopher and Fredrik Sävje and Daniel Spielman and Peng Zhang (2019),
   \sQuote{Balancing covariates in randomized experiments using the Gram-Schmidt walk}, arXiv 1911.03071.
   \url{https://arxiv.org/abs/1911.03071}
}
