% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gswdesign-package.R
\name{gswdesign_setup}
\alias{gswdesign_setup}
\title{Initial setup for the gswdesign package.}
\usage{
gswdesign_setup(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[JuliaCall]{julia_setup}}.}
}
\description{
\code{gswdesign_setup} does the initial setup for the gswdesign package.
It should be run before calling any of the functions in the package.
}
\examples{
\dontrun{
gswdesign_setup()
}

}
