% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gswdesign-package.R
\name{gswdesign_update}
\alias{gswdesign_update}
\title{Update GSWDesign package in Julia}
\usage{
gswdesign_update()
}
\description{
\code{gswdesign_update} updates the GSWDesign Julia package.
}
\examples{
\dontrun{
gswdesign_setup()
gswdesign_update()
}

}
