% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gswdesign-package.R
\name{julia_running}
\alias{julia_running}
\title{Check so Julia is setup and running}
\usage{
julia_running()
}
\description{
\code{julia_running} checks so Julia is running and can be called.
}
\examples{
\dontrun{
julia_running()
}

}
