% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gsw.R
\name{set_julia_seed}
\alias{set_julia_seed}
\title{Set RNG seed in Julia}
\usage{
set_julia_seed(seed)
}
\arguments{
\item{seed}{the seed to use}
}
\description{
Sets the seed for the random number generator in Julia.
}
\examples{
\dontrun{
# Set Julia seed to "12345"
set_julia_seed(12345)
}

}
