% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gswdesign-package.R
\docType{package}
\name{gswdesign-package}
\alias{gswdesign-package}
\title{gswdesign: The Gram-Schmidt walk design}
\description{
Provides wrappers for the Julia package 'GSWDesign', which is a fast
implementation of the Gram--Schmidt Walk for balancing covariates in
randomized experiments. The Gram--Schmidt Walk design allows experimenters
the flexibilty to control the amount of covariate balancing.
}
\details{
See \code{\link{sample_gs_walk}} for the main function.

See the packages' websites for more information:
\url{https://github.com/crharshaw/GSWDesign.jl},
\url{https://github.com/fsavje/gswdesign-R}.

Bug reports and suggestions are greatly appreciated. They are best reported
here: \url{https://github.com/fsavje/gswdesign-R/issues}.
}
\references{
Harshaw, Christopher and Fredrik Sävje and Daniel Spielman and Peng Zhang (2019),
   \sQuote{Balancing covariates in randomized experiments using the Gram-Schmidt walk}, arXiv 1911.03071.
   \url{https://arxiv.org/abs/1911.03071}
}
