% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gsw.R
\name{sample_gsw}
\alias{sample_gsw}
\title{Sample from the Gram-Schmidt Walk design.}
\usage{
sample_gsw(X, lambda = 0.5, verbose = FALSE, balanced = FALSE)
}
\arguments{
\item{X}{a n-by-x matrix with covariates to balance}

\item{lambda}{a real value in (0,1) specifying weight}

\item{verbose}{set to true to print out algorithm state}

\item{balanced}{set true to run the balanced GSW}
}
\description{
An fast implementation for sampling from the Gram-Schmidt Walk Design.
Maintains a cholesky factorization of (I + A * A^T ) for faster repeated linear
system solves and has a recursive component for more effective memory allocation.
}
\examples{
\dontrun{
gswdesign_setup()
X <- matrix(rnorm(1000), nrow = 100)
sample_gsw(X, lambda = 0.3)
}

}
